requirements.txt
spade>=4.1.2
jsonschema>=4.20.0
aiohttp>=3.9
pyyaml>=6.0
pytest>=8.0
pytest-aiohttp>=1.0
pytest-asyncio>=0.23
pytest-timeout>=2.2
beautifulsoup4>=4.12
lxml>=5.1
